# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Open Archiver API",
    "version": "1.0.0",
    "description": "REST API for Open Archiver — an open-source email archiving platform. All authenticated endpoints require a Bearer JWT token obtained from `POST /v1/auth/login`, or an API key passed as a Bearer token.",
    "license": {
      "name": "SEE LICENSE IN LICENSE",
      "url": "https://github.com/LogicLabs-OU/OpenArchiver/blob/main/LICENSE"
    },
    "contact": {
      "name": "Open Archiver",
      "url": "https://openarchiver.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    },
    {
      "apiKeyAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT obtained from `POST /v1/auth/login`. Pass as `Authorization: Bearer <token>`."
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-KEY",
        "description": "API key generated via `POST /v1/api-keys`. Pass as `X-API-KEY: <key>`."
      }
    },
...
    "schemas": {
      "ErrorMessage": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Human-readable error description.",
            "example": "An error occurred."
          }
        },
        "required": [
          "message"
        ]
      },
...
      "SafeIngestionSource": {
        "type": "object",
        "description": "An ingestion source with sensitive credential fields removed.",
        "properties": {
          "id": {
            "type": "string",
            "example": "clx1y2z3a0000b4d2"
          },
          "name": {
            "type": "string",
            "example": "Company Google Workspace"
          },
          "provider": {
            "type": "string",
            "enum": [
              "google_workspace",
              "microsoft_365",
              "generic_imap",
              "pst_import",
              "eml_import",
              "mbox_import"
            ],
            "example": "google_workspace"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "paused",
              "error",
              "pending_auth",
              "syncing",
              "importing",
              "auth_success",
              "imported"
            ],
            "example": "active"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastSyncStartedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lastSyncFinishedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lastSyncStatusMessage": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "name",
          "provider",
          "status",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateIngestionSourceDto": {
        "type": "object",
        "required": [
          "name",
          "provider",
          "providerConfig"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "Company Google Workspace"
          },
          "provider": {
            "type": "string",
            "enum": [
              "google_workspace",
              "microsoft_365",
              "generic_imap",
              "pst_import",
              "eml_import",
              "mbox_import"
            ]
          },
          "providerConfig": {
            "type": "object",
            "description": "Provider-specific configuration. See the ingestion source guides for the required fields per provider.",
            "example": {
              "serviceAccountKeyJson": "{\"type\":\"service_account\",...}",
              "impersonatedAdminEmail": "admin@example.com"
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/v1/ingestion-sources": {
      "post": {
        "summary": "Create an ingestion source",
        "description": "Creates a new ingestion source and validates the connection. Returns the created source without credentials. Requires `create:ingestion` permission.",
        "operationId": "createIngestionSource",
        "tags": [
          "Ingestion"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateIngestionSourceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Ingestion source created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeIngestionSource"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or connection test failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "get": {
        "summary": "List ingestion sources",
        "description": "Returns all ingestion sources accessible to the authenticated user. Credentials are excluded from the response. Requires `read:ingestion` permission.",
        "operationId": "listIngestionSources",
        "tags": [
          "Ingestion"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Array of ingestion sources.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SafeIngestionSource"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/v1/ingestion-sources/reindex-all": {
      "post": {
        "summary": "Reindex the entire archive",
        "description": "Enqueues a reindex of every ingestion source. Rebuilds search documents from existing archived emails without re-ingesting. Requires `manage:ingestion` permission.",
        "operationId": "reindexAllIngestionSources",
        "tags": [
          "Ingestion"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "mode": {
                    "type": "string",
                    "enum": [
                      "missing",
                      "full"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Reindex job accepted and queued."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/v1/ingestion-sources/{id}": {
      "get": {
        "summary": "Get an ingestion source",
        "description": "Returns a single ingestion source by ID. Credentials are excluded. Requires `read:ingestion` permission.",
        "operationId": "getIngestionSourceById",
        "tags": [
          "Ingestion"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "clx1y2z3a0000b4d2"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ingestion source details.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeIngestionSource"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "put": {
        "summary": "Update an ingestion source",
        "description": "Updates configuration for an existing ingestion source. Requires `update:ingestion` permission.",
        "operationId": "updateIngestionSource",
        "tags": [
          "Ingestion"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "clx1y2z3a0000b4d2"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateIngestionSourceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated ingestion source.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeIngestionSource"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "delete": {
        "summary": "Delete an ingestion source",
        "description": "Permanently deletes an ingestion source. Deletion must be enabled in system settings. Requires `delete:ingestion` permission.",
        "operationId": "deleteIngestionSource",
        "tags": [
          "Ingestion"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "clx1y2z3a0000b4d2"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Ingestion source deleted. No content returned."
          },
          "400": {
            "description": "Deletion disabled or constraint error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/v1/ingestion-sources/{id}/import": {
      "post": {
        "summary": "Trigger initial import",
        "description": "Enqueues an initial import job for the ingestion source. This imports all historical emails. Requires `create:ingestion` permission.",
        "operationId": "triggerInitialImport",
        "tags": [
          "Ingestion"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "clx1y2z3a0000b4d2"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Initial import job accepted and queued.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/v1/ingestion-sources/{id}/pause": {
      "post": {
        "summary": "Pause an ingestion source",
        "description": "Sets the ingestion source status to `paused`, stopping continuous sync. Requires `update:ingestion` permission.",
        "operationId": "pauseIngestionSource",
        "tags": [
          "Ingestion"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "clx1y2z3a0000b4d2"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ingestion source paused. Returns the updated source.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeIngestionSource"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/v1/ingestion-sources/{id}/sync": {
      "post": {
        "summary": "Force sync",
        "description": "Triggers an out-of-schedule continuous sync for the ingestion source. Requires `sync:ingestion` permission.",
        "operationId": "triggerForceSync",
        "tags": [
          "Ingestion"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "clx1y2z3a0000b4d2"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Force sync job accepted and queued.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/v1/ingestion-sources/{id}/reindex": {
      "post": {
        "summary": "Reindex an ingestion source",
        "description": "Enqueues a reindex of the source (and its merge group). Rebuilds search documents from existing archived emails without re-ingesting or duplicating. Requires `sync:ingestion` permission.",
        "operationId": "reindexIngestionSource",
        "tags": [
          "Ingestion"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "mode": {
                    "type": "string",
                    "enum": [
                      "missing",
                      "full"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Reindex job accepted and queued."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/v1/ingestion-sources/{id}/stats": {
      "get": {
        "summary": "Get statistics for a source",
        "description": "Returns read-only statistics for the source (and its merge group) — email/mailbox/thread counts, storage usage, index coverage, attachment/compliance counts, per-mailbox breakdown, merge-group children, and recent activity. Requires `read:ingestion` permission.",
        "operationId": "getIngestionSourceStats",
        "tags": [
          "Ingestion"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ingestion source statistics."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/v1/ingestion-sources/{id}/unmerge": {
      "post": {
        "summary": "Unmerge a child ingestion source",
        "description": "Detaches a child source from its merge group, making it a standalone root source. Requires `update:ingestion` permission.",
        "operationId": "unmergeIngestionSource",
        "tags": [
          "Ingestion"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Source unmerged. Returns the updated source.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeIngestionSource"
                }
              }
            }
          },
          "400": {
            "description": "Source is not merged into another source."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
...
}